+++ example/a.ts
@@ -1,27 +1,25 @@
 import { Elysia, t } from '../src'
+import { req } from '../test/utils'
 
-export const IMPORT_SCHEMA = t.Object({
-	file: t.File({ error: 'File tidak valid', type: ['text/csv'] })
-})
-
-export const IMPORT_DETAIL = {
-	summary: 'Import',
-	description: 'Impor data pengguna dari file CSV'
-}
-
-new Elysia()
-	.post(
-		'/import',
-		({ body }) => {
-			const { file } = body
-			console.debug('Uploaded file: ', file)
-			return {
-				message: 'File upload'
-			}
+const app = new Elysia()
+	.model({
+		a: t.String(),
+		myModel: t.Object({ num: t.Number(), a: t.Ref('a') })
+	})
+	.get(
+		'/',
+		({ query }) => {
+			return query
 		},
 		{
-			body: IMPORT_SCHEMA,
-			detail: IMPORT_DETAIL
+			query: 'myModel'
 		}
 	)
-	.listen(3000)
+
+app.handle(req('/?num=1&a=a'))
+	.then((x) => x.text())
+	.then(console.log)
+
+console.log(
+	`🦊 Elysia is running at ${app.server?.hostname}:${app.server?.port}`
+)